+++ src/main/java/com/kdt/config/WebSocketConfig.java
@@ -1,0 +1,16 @@
+//package com.kdt.config;
+//
+//import org.springframework.context.annotation.Configuration;
+//import org.springframework.web.socket.config.annotation.EnableWebSocket;
+//import org.springframework.web.socket.config.annotation.WebSocketConfigurer;
+//import org.springframework.web.socket.config.annotation.WebSocketHandlerRegistry;
+//
+//@Configuration
+//@EnableWebSocket
+//public class WebSocketConfig implements WebSocketConfigurer  {
+//
+////	@Override
+////    public void registerWebSocketHandlers(WebSocketHandlerRegistry registry) {
+////        registry.addHandler(new YourWebSocketHandler(), "/ws").setAllowedOrigins("*");
+////    }
+//}
