+++ src/main/java/com/kdt/controllers/WebSocketController.java
@@ -1,0 +1,8 @@
+package com.kdt.controllers;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class WebSocketController {
+
+}
